+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - [ ] When the user presses enter it may remove all the text (typewriter) and then do something.
 - [ ] FIX: the logo in the navbar.
 - [ ] When switching different tabs it should have a swoosh (slide left) animation or side rigth if reverse
+- [ ] Make the terminal interactive
 
 ## CHANGELOGS: